requirements.txt
asgiref==3.7.2
Brotli==1.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.5
django-crispy-forms==2.0
filelock==3.12.4
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.1
platformdirs==3.10.0
protobuf==4.24.4
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
virtualenv==20.24.5
whitenoise==6.6.0
